perbaikan navbar menjadi component terpisah dengan app
